Cosas de brandon

+++ app/src/main/java/com/example/purchaez/recycleView/Items.kt
@@ -3,9 +3,9 @@ package com.example.purchaez.recycleView
 class Items{
 
     var name: String = ""
-    var cost: Int = 0
+    var cost: Float = 0.0F
 
-    constructor(newName:String, newCost:Int){
+    constructor(newName:String, newCost:Float){
         name = newName
         cost = newCost
     }
